requirements.txt
amqp==5.0.9
asgiref==3.4.1
billiard==3.6.4.0
celery==5.2.3
click==8.0.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
Django==4.0.1
django-admin-tools==0.9.2
django-filter==21.1
djangorestframework==3.13.1
Faker==11.3.0
feedparser==6.0.8
importlib-metadata==4.10.1
kombu==5.2.3
Markdown==3.3.6
mixer==7.2.0
prompt-toolkit==3.0.24
python-dateutil==2.8.2
pytz==2021.3
sgmllib3k==1.0.0
six==1.16.0
sqlparse==0.4.2
text-unidecode==1.3
vine==5.0.0
wcwidth==0.2.5
zipp==3.7.0
